REFACTOR: Move domain to "auth" from "user"
